Added api to update and insert insurance, amc, warranty and puc
Fixed api for deleting image from product item and mailbox.

Updated code to fix if product don't have job linked

Updated code of adding product item with new effective date

Added extended provider name

+++ src/api/Adaptors/amcs.js
@@ -310,12 +310,26 @@ class AmcAdaptor {
   }
 
   deleteAMC(id, user_id) {
-    return this.modals.amcs.destroy({
-      where: {
-        id,
-        user_id,
-      },
-    }).then(() => {
+    return this.modals.amcs.findById(id).then((result) => {
+      if (result) {
+        return Promise.all([
+          this.modals.amcs.destroy({
+            where: {
+              id,
+              user_id,
+            },
+          }), result.copies.length > 0 ? this.modals.jobCopies.update({
+            status_type: 3,
+            updated_by: user_id,
+          }, {
+            where: {
+              id: result.copies.map(item => item.copyId),
+            },
+          }) : undefined]).then(() => {
+          return true;
+        });
+      }
+
       return true;
     });
   }
