+++ src/api/Adaptors/insurances.js
@@ -382,12 +382,26 @@ class InsuranceAdaptor {
   }
 
   deleteInsurance(id, user_id) {
-    return this.modals.insurances.destroy({
-      where: {
-        id,
-        user_id,
-      },
-    }).then(() => {
+    return this.modals.insurances.findById(id).then((result) => {
+      if (result) {
+        return Promise.all([
+          this.modals.insurances.destroy({
+            where: {
+              id,
+              user_id,
+            },
+          }), result.copies.length > 0 ? this.modals.jobCopies.update({
+            status_type: 3,
+            updated_by: user_id,
+          }, {
+            where: {
+              id: result.copies.map(item => item.copyId),
+            },
+          }) : undefined]).then(() => {
+          return true;
+        });
+      }
+
       return true;
     });
   }
